+++ packages/modal/src/components/content/Component.tsx
@@ -10,13 +10,14 @@ import styles from './index.module.css';
 import mobileStyles from './mobile.module.css';
 
 export const Content: FC<ContentProps> = ({ children, flex, className }) => {
-    const { contentRef } = useContext(ModalContext);
+    const { contentRef, hasHeader } = useContext(ModalContext);
     const { size, view } = useContext(ResponsiveContext);
 
     return (
         <div
             className={cn(styles.content, className, {
                 [styles.flex]: flex,
+                [styles.withHeader]: hasHeader,
                 [desktopStyles[size]]: view === 'desktop' && size,
                 [mobileStyles.content]: view === 'mobile',
             })}