+++ packages/comment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.1.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@3.2.0
+
 ## 2.1.8
 
 ### Patch Changes