+++ packages/select/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 13.2.13
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   bottom-sheet@6.0.0
+    -   modal@7.0.0
+    -   base-modal@5.1.0
+
 ## 13.2.12
 
 ### Patch Changes